Fixed pipeline vs. VBO
Since I discovered annoying flickerings on NVIDIA's Linux drivers, I decided to change the rendering subsystem to modern OpenGL 3.1 using VBO and shaders. And yeah, the flickerings are gone. But then I experienced a massive performance drop. This is because updating and calling VBO for each tile. Per frame, approx. 1000-1500 tiles are painted. After reading myriad of articles how to speed up vbo streaming, I reimplemented the old-fashioned fixed pipeline rendering and I offered a new option to toggle between both options. Sorry for that. In future, I really want to remove the old glBegin()/glEnd() rendering technique to take advantage of shaders on the GPU. But for now, we have to live with that. If you have really nice vbo streaming optimization hints, let me know or implement it yourself.