improved the lagometer
the lagometer shows the ping (green-red), the server outqueue stack size (yellow) and the client outqueue stack size (blue). ping should be <50 (optimal gameplay). the yellow bar should not increase. if that happens, it indicates, that the server cannot send data in time anymore. the blue bar indicates the same behavior from the perspective of the client.